+++ glyph-brush/CHANGELOG.md
@@ -1,4 +1,7 @@
 # Unreleased (v0.7.7)
+* Allow `Text::new` to work with any `X` type. **This may break usage**, however it will hopefully be non-breaking in practice as the compiler should always be able to infer this.
+* Add `Section::builder` for more flexible `X`/"extra" type usage than `Section::default`.
+* Add more flexible `X` type usage to `GlyphBrush::keep_cached`.
 * Update `GlyphCruncher::glyphs`, `GlyphCruncher::glyph_bounds` docs.
 
 # v0.7.6